Changed Patch model to use a long for the baseFileVersion

+++ src/main/java/patching/Patch.java
@@ -9,10 +9,10 @@ import java.util.List;
 
 public class Patch {
 
-    private final int baseVersion;
+    private final long baseVersion;
     private final List<Diff> diffs;
 
-    public Patch(int baseVersion, List<Diff> diffs) {
+    public Patch(long baseVersion, List<Diff> diffs) {
         this.baseVersion = baseVersion;
         this.diffs = diffs;
     }
@@ -33,7 +33,7 @@ public class Patch {
         return diffs;
     }
 
-    public int getBaseVersion() {
+    public long getBaseVersion() {
         return baseVersion;
     }
 
@@ -83,7 +83,7 @@ public class Patch {
     public int hashCode() {
         final int prime = 31;
         int result = 1;
-        result = prime * result + baseVersion;
+        result = prime * result + (int) baseVersion;
         result = prime * result + ((diffs == null) ? 0 : diffs.hashCode());
         return result;
     }
@@ -113,7 +113,7 @@ public class Patch {
 
     public Patch transform(Patch... patches) {
         List<Diff> intermediateDiffs = diffs;
-        int maxVersionSeen = baseVersion;
+        long maxVersionSeen = baseVersion;
 
         for (Patch patch : patches) {
             // Must be able to transform backwards as well?
